added command line support

+++ script/bakapk.sh
@@ -1,29 +1,36 @@
 #!/bin/sh
 
-(cd ..
-    folder="${1%.*}"
-    folder=my$folder
-    dest=/tmp/
+if [ $# != 1 ]
+then
+	echo "Usage: $0 name.apk(under test/apk/)"
+else
+	(cd ..
+	folder="${1%.*}"
+	folder=my$folder
+	dest=/tmp/$folder
 
-    rm -r smalioutput 2>/dev/null
-    rm -r output 2>/dev/null
-    rm -r /tmp/output 2>/dev/null
-    rm -r $dest$folder 2>/dev/null
+	rm -r /tmp/output 2>/dev/null
+	rm -r /tmp/smalioutput 2>/dev/null
+	rm -r $dest 2>/dev/null
 
-    export CLASSPATH="."
-    for i in jar/*.jar
-    do
-        export CLASSPATH="$CLASSPATH":$i
-    done
+	export CLASSPATH="."
+	for i in jar/*.jar
+	do
+		export CLASSPATH="$CLASSPATH":$i
+	done
 
-    # will generate output/ and smalioutput/
-    java -Xmx2g -Xms2g -cp .:./bin:$CLASSPATH Carbon ./test/apk/$1
+	# will generate /tmp/output/ and /tmp/smalioutput/
+	java -Xmx2g -Xms2g -cp .:./bin:$CLASSPATH Carbon ./test/apk/$1
 
-    cp -r output /tmp/
-    mv  output $dest$folder
-    rm -r  $dest$folder/smali
-    mv smalioutput $dest$folder/smali
-    java -jar jar/apktool.jar b $dest$folder
+	mkdir $dest
+	# only copy needed file, ignore /tmp/output/smali
+	for i in `find /tmp/output/ -maxdepth 1|grep -v smali| tail -n +2`
+	do
+		cp -r $i $dest
+	done
+	mv /tmp/smalioutput $dest/smali
+	java -jar jar/apktool.jar b $dest
 
-    java -jar jar/SignAPK.jar $dest$folder/dist/$1  $dest${folder}.apk
-)
+	java -jar jar/SignAPK.jar $dest/dist/$1 $dest.apk
+	)
+fi
